move the other job subfunctions to the job file
