+++ operator/pkg/upgrade/upgrade.go
@@ -6,35 +6,24 @@ import (
 	"strings"
 	"time"
 
-	"github.com/google/uuid"
 	apv1b2 "github.com/k0sproject/k0s/pkg/apis/autopilot/v1beta2"
-	autopilotv1beta2 "github.com/k0sproject/k0s/pkg/apis/autopilot/v1beta2"
 	k0shelm "github.com/k0sproject/k0s/pkg/apis/helm/v1beta1"
 	k0sv1beta1 "github.com/k0sproject/k0s/pkg/apis/k0s/v1beta1"
 	"github.com/replicatedhq/embedded-cluster/kinds/apis/v1beta1"
 	clusterv1beta1 "github.com/replicatedhq/embedded-cluster/kinds/apis/v1beta1"
-	"github.com/replicatedhq/embedded-cluster/operator/pkg/artifacts"
 	"github.com/replicatedhq/embedded-cluster/operator/pkg/autopilot"
 	"github.com/replicatedhq/embedded-cluster/operator/pkg/charts"
 	"github.com/replicatedhq/embedded-cluster/operator/pkg/k8sutil"
 	"github.com/replicatedhq/embedded-cluster/operator/pkg/release"
-	batchv1 "k8s.io/api/batch/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
 const (
-	operatorChartName      = "embedded-cluster-operator"
-	clusterConfigName      = "k0s"
-	clusterConfigNamespace = "kube-system"
-	upgradeJobName         = "embedded-cluster-upgrade-%s"
-	upgradeJobConfigMap    = "upgrade-job-configmap-%s"
+	operatorChartName   = "embedded-cluster-operator"
+	upgradeJobName      = "embedded-cluster-upgrade-%s"
+	upgradeJobConfigMap = "upgrade-job-configmap-%s"
 )
 
 // Upgrade upgrades the embedded cluster to the version specified in the installation.
@@ -224,38 +213,6 @@ func chartUpgrade(ctx context.Context, cli client.Client, in *clusterv1beta1.Ins
 
 }
 
-func createInstallation(ctx context.Context, cli client.Client, original *clusterv1beta1.Installation) error {
-	log := ctrl.LoggerFrom(ctx)
-	in := original.DeepCopy()
-
-	// check if the installation already exists - this function can be called multiple times
-	// if the installation is already created, we can just return
-	nsn := types.NamespacedName{Name: in.Name}
-	var existingInstallation clusterv1beta1.Installation
-	if err := cli.Get(ctx, nsn, &existingInstallation); err == nil {
-		log.Info(fmt.Sprintf("Installation %s already exists", in.Name))
-		return nil
-	}
-	log.Info(fmt.Sprintf("Creating installation %s", in.Name))
-
-	err := cli.Create(ctx, in)
-	if err != nil {
-		return fmt.Errorf("create installation: %w", err)
-	}
-
-	// set the state to 'waiting' so that the operator will not reconcile based on it
-	// we will set the state to installed after the installation is complete
-	in.Status.State = v1beta1.InstallationStateWaiting
-	err = cli.Status().Update(ctx, in)
-	if err != nil {
-		return fmt.Errorf("update installation status: %w", err)
-	}
-
-	log.Info("Installation created")
-
-	return nil
-}
-
 func unLockInstallation(ctx context.Context, cli client.Client, in *v1beta1.Installation) error {
 	existingInstallation := &v1beta1.Installation{}
 	err := cli.Get(ctx, client.ObjectKey{Name: in.Name}, existingInstallation)
@@ -289,175 +246,4 @@ func waitForOperatorChart(ctx context.Context, cli client.Client, version string
 		return ready, nil
 	})
 	return err
-}
-
-const (
-	installationNameAnnotation = "embedded-cluster.replicated.com/installation-name"
-)
-
-func airgapDistributeArtifacts(ctx context.Context, cli client.Client, in *clusterv1beta1.Installation, localArtifactMirrorImage string) error {
-	// in airgap installations let's make sure all assets have been copied to nodes.
-	// this may take some time so we only move forward when 'ready'.
-	err := ensureAirgapArtifactsOnNodes(ctx, cli, in, localArtifactMirrorImage)
-	if err != nil {
-		return fmt.Errorf("ensure airgap artifacts: %w", err)
-	}
-
-	// once all assets are in place we can create the autopilot plan to push the images to
-	// containerd.
-	err = ensureAirgapArtifactsInCluster(ctx, cli, in)
-	if err != nil {
-		return fmt.Errorf("autopilot copy airgap artifacts: %w", err)
-	}
-
-	return nil
-}
-
-func ensureAirgapArtifactsOnNodes(ctx context.Context, cli client.Client, in *clusterv1beta1.Installation, localArtifactMirrorImage string) error {
-	log := ctrl.LoggerFrom(ctx)
-
-	log.Info("Placing artifacts on nodes...")
-
-	op, err := artifacts.EnsureRegistrySecretInECNamespace(ctx, cli, in)
-	if err != nil {
-		return fmt.Errorf("ensure registry secret in ec namespace: %w", err)
-	} else if op != controllerutil.OperationResultNone {
-		log.Info("Registry credentials secret changed", "operation", op)
-	}
-
-	err = artifacts.EnsureArtifactsJobForNodes(ctx, cli, in, localArtifactMirrorImage)
-	if err != nil {
-		return fmt.Errorf("ensure artifacts job for nodes: %w", err)
-	}
-
-	log.Info("Waiting for artifacts to be placed on nodes...")
-
-	err = wait.PollUntilContextCancel(ctx, 5*time.Second, true, func(ctx context.Context) (bool, error) {
-		jobs, err := artifacts.ListArtifactsJobForNodes(ctx, cli, in)
-		if err != nil {
-			return false, fmt.Errorf("list artifacts jobs for nodes: %w", err)
-		}
-
-		ready := true
-		for nodeName, job := range jobs {
-			if job == nil {
-				return false, fmt.Errorf("job for node %s not found", nodeName)
-			}
-			if job.Status.Succeeded > 0 {
-				continue
-			}
-			ready = false
-			for _, cond := range job.Status.Conditions {
-				if cond.Type == batchv1.JobFailed {
-					if cond.Status == corev1.ConditionTrue {
-						// fail immediately if any job fails
-						return false, fmt.Errorf("job for node %s failed: %s - %s", nodeName, cond.Reason, cond.Message)
-					}
-					break
-				}
-			}
-			// job is still running
-		}
-
-		return ready, nil
-	})
-	if err != nil {
-		return fmt.Errorf("wait for artifacts job for nodes: %w", err)
-	}
-
-	log.Info("Artifacts placed on nodes")
-	return nil
-}
-
-func ensureAirgapArtifactsInCluster(ctx context.Context, cli client.Client, in *clusterv1beta1.Installation) error {
-	log := ctrl.LoggerFrom(ctx)
-
-	log.Info("Uploading container images...")
-
-	err := autopilotEnsureAirgapArtifactsPlan(ctx, cli, in)
-	if err != nil {
-		return fmt.Errorf("ensure autopilot plan: %w", err)
-	}
-
-	nsn := types.NamespacedName{Name: "autopilot"}
-	plan := autopilotv1beta2.Plan{}
-
-	log.Info("Waiting for container images to be uploaded...")
-
-	err = wait.PollUntilContextCancel(ctx, 5*time.Second, true, func(ctx context.Context) (bool, error) {
-		err := cli.Get(ctx, nsn, &plan)
-		if err != nil {
-			return false, fmt.Errorf("get autopilot plan: %w", err)
-		}
-		if plan.Annotations[installationNameAnnotation] != in.Name {
-			return false, fmt.Errorf("autopilot plan for different installation")
-		}
-
-		switch {
-		case autopilot.HasPlanSucceeded(plan):
-			return true, nil
-		case autopilot.HasPlanFailed(plan):
-			reason := autopilot.ReasonForState(plan)
-			return false, fmt.Errorf("autopilot plan failed: %s", reason)
-		}
-		// plan is still running
-		return false, nil
-	})
-	if err != nil {
-		return fmt.Errorf("wait for autopilot plan: %w", err)
-	}
-
-	log.Info("Container images uploaded")
-	return nil
-}
-
-func autopilotEnsureAirgapArtifactsPlan(ctx context.Context, cli client.Client, in *clusterv1beta1.Installation) error {
-	plan, err := getAutopilotAirgapArtifactsPlan(ctx, cli, in)
-	if err != nil {
-		return fmt.Errorf("get autopilot airgap artifacts plan: %w", err)
-	}
-
-	err = k8sutil.EnsureObject(ctx, cli, plan, func(opts *k8sutil.EnsureObjectOptions) {
-		opts.ShouldDelete = func(obj client.Object) bool {
-			return obj.GetAnnotations()[installationNameAnnotation] != in.Name
-		}
-	})
-	if err != nil {
-		return fmt.Errorf("ensure autopilot plan: %w", err)
-	}
-
-	return nil
-}
-
-func getAutopilotAirgapArtifactsPlan(ctx context.Context, cli client.Client, in *clusterv1beta1.Installation) (*autopilotv1beta2.Plan, error) {
-	var commands []autopilotv1beta2.PlanCommand
-
-	// if we are running in an airgap environment all assets are already present in the
-	// node and are served by the local-artifact-mirror binary listening on localhost
-	// port 50000. we just need to get autopilot to fetch the k0s binary from there.
-	command, err := artifacts.CreateAutopilotAirgapPlanCommand(ctx, cli, in)
-	if err != nil {
-		return nil, fmt.Errorf("create autopilot airgap plan command: %w", err)
-	}
-	commands = append(commands, *command)
-
-	plan := &autopilotv1beta2.Plan{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: autopilotv1beta2.SchemeGroupVersion.String(),
-			Kind:       "Plan",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "autopilot", // this is a fixed name and should not be changed
-			Annotations: map[string]string{
-				installationNameAnnotation: in.Name,
-			},
-		},
-		Spec: autopilotv1beta2.PlanSpec{
-			Timestamp: "now",
-			ID:        uuid.New().String(),
-			Commands:  commands,
-		},
-	}
-
-	return plan, nil
 }
